Ensure containers with static IP addresses have DNS info
Since the switch to generate unique MAAS hostnames and with it a switch
to use static IP address allocation a container now has an empty
/etc/resolv.conf; its DNS info previously came from the DHCP lease.

This change augments the manually configured interface info by using the
existing code that parses /etc/resolv.conf (on the host) and adds that
to the containers' interface information.

This change deliberately does not contain additional unit tests for this
case, it is here as the smallest change possible and is to be used for
unblocking testing. If the change is good, a follow-up PR will include
unit tests.

Fixes [LP:#1528217](https://bugs.launchpad.net/juju-core/+bug/1528217)